requirements.txt
selenium>=4.0.0
pytest>=7.0.0
allure-pytest>=2.9.0
configparser>=5.0.0
webdriver-manager>=3.0.0